Found a way to render popup.html for without clicking on it
This is a hack to get the HTML from the renderer popup.html as if the
used has clicked on it. Controlling - even in end to end tests - is not
possible.
The chromium team has also stated that it will never be possible to
"click" on the browserAction with JS (which is a good secure thing to
do).

So we create an iframe with the popup URL in the background page.
Since the JS needed by the popup is available and the macthing rules are
already found the iframe will render the popup as if the user clicked
the browserAction.

voilá